Refactor agent API
