+++ server/modules/routers.js
@@ -5,6 +5,7 @@ import AuthRouter from './auth/auth.route';
 import TicketRouter from './ticket/ticket.route';
 import ChatLogRouter from './chatlog/chatlog.route';
 import ApplicationRouter from './application/application.route';
+import AgentRouter from './agent/agent.route';
 
 const router = express.Router();
 
@@ -14,5 +15,6 @@ export default () => {
   router.use('/chats', ChatLogRouter.router);
   router.use('/tickets', TicketRouter.router);
   router.use('/applications', ApplicationRouter.router);
+  router.use('/agents', AgentRouter.router);
   return router;
 };
